Код
- Начал реализовывать протокол верхнего уровня. На данный момент PC формирует UDP пакет и передает его узлу. Узел обрабатывает весь стек от eth->route->ip->udp->port handler. Но есть какая то ошибка, второй раз не вызывается и я видел что дебагер ругнулся на размер стека. Нужно отлаживать дальше

+++ IAR/NeocoreGateway/src/cmd_0x01.c
@@ -7,7 +7,7 @@
 #define ARGS_SIZE sizeof(cmd_args_s)
 typedef struct //!< Аргументы команды
 {
-  uint16_t panid;
+  uint8_t panid;
   uint16_t crc16;
 } cmd_args_s;
